refactor: separate all graphql definitions info files
