+++ src/gql/mutations/EditUserInfoMutation.ts
@@ -1,0 +1,11 @@
+import { graphql } from 'babel-plugin-relay/macro'
+
+const EditUserInfoMutation = graphql`
+  mutation EditUserInfoMutation($name: String, $age: Int, $bio: String) {
+    editUser(name: $name, age: $age, bio: $bio) {
+      id
+    }
+  }
+`
+
+export default EditUserInfoMutation
