+++ src/gql/query/NotesListQuery.ts
@@ -1,0 +1,13 @@
+import { graphql } from 'babel-plugin-relay/macro'
+
+const NotesListQuery = graphql`
+  query NotesListQuery {
+    notes {
+      id
+      title
+      description
+    }
+  }
+`
+
+export default NotesListQuery
