+++ src/gql/query/UserInfoQuery.ts
@@ -1,0 +1,14 @@
+import { graphql } from 'babel-plugin-relay/macro'
+
+const UserInfoQuery = graphql`
+  query UserInfoQuery {
+    user {
+      id
+      name
+      email
+      bio
+      age
+    }
+  }
+`
+export default UserInfoQuery
